refactor util (#2259)
* refactor util

* merge common with ops
