+++ mmdet/models/anchor_heads/atss_head.py
@@ -7,9 +7,10 @@ from mmcv.cnn import normal_init
 from mmdet.core import (PseudoSampler, anchor_inside_flags, bbox2delta,
                         build_assigner, delta2bbox, force_fp32,
                         images_to_levels, multi_apply, multiclass_nms, unmap)
+from mmdet.ops import ConvModule, Scale
 from ..builder import build_loss
 from ..registry import HEADS
-from ..utils import ConvModule, Scale, bias_init_with_prob
+from ..utils import bias_init_with_prob
 from .anchor_head import AnchorHead
