+++ mmdet/models/anchor_heads/fcos_head.py
@@ -3,9 +3,10 @@ import torch.nn as nn
 from mmcv.cnn import normal_init
 
 from mmdet.core import distance2bbox, force_fp32, multi_apply, multiclass_nms
+from mmdet.ops import ConvModule, Scale
 from ..builder import build_loss
 from ..registry import HEADS
-from ..utils import ConvModule, Scale, bias_init_with_prob
+from ..utils import bias_init_with_prob
 
 INF = 1e8
