+++ mmdet/models/anchor_heads/fovea_head.py
@@ -3,10 +3,10 @@ import torch.nn as nn
 from mmcv.cnn import normal_init
 
 from mmdet.core import multi_apply, multiclass_nms
-from mmdet.ops import DeformConv
+from mmdet.ops import ConvModule, DeformConv
 from ..builder import build_loss
 from ..registry import HEADS
-from ..utils import ConvModule, bias_init_with_prob
+from ..utils import bias_init_with_prob
 
 INF = 1e8
