+++ mmdet/models/anchor_heads/ga_retina_head.py
@@ -1,9 +1,9 @@
 import torch.nn as nn
 from mmcv.cnn import normal_init
 
-from mmdet.ops import MaskedConv2d
+from mmdet.ops import ConvModule, MaskedConv2d
 from ..registry import HEADS
-from ..utils import ConvModule, bias_init_with_prob
+from ..utils import bias_init_with_prob
 from .guided_anchor_head import FeatureAdaption, GuidedAnchorHead
