+++ mmdet/models/anchor_heads/reppoints_head.py
@@ -7,10 +7,10 @@ from mmcv.cnn import normal_init
 
 from mmdet.core import (PointGenerator, multi_apply, multiclass_nms,
                         point_target)
-from mmdet.ops import DeformConv
+from mmdet.ops import ConvModule, DeformConv
 from ..builder import build_loss
 from ..registry import HEADS
-from ..utils import ConvModule, bias_init_with_prob
+from ..utils import bias_init_with_prob
 
 
 @HEADS.register_module
