+++ mmdet/models/anchor_heads/retina_sepbn_head.py
@@ -2,8 +2,9 @@ import numpy as np
 import torch.nn as nn
 from mmcv.cnn import normal_init
 
+from mmdet.ops import ConvModule
 from ..registry import HEADS
-from ..utils import ConvModule, bias_init_with_prob
+from ..utils import bias_init_with_prob
 from .anchor_head import AnchorHead
