+++ mmdet/models/backbones/resnext.py
@@ -2,8 +2,8 @@ import math
 
 import torch.nn as nn
 
+from mmdet.ops import build_conv_layer, build_norm_layer
 from ..registry import BACKBONES
-from ..utils import build_conv_layer, build_norm_layer
 from .resnet import Bottleneck as _Bottleneck
 from .resnet import ResNet
