+++ mmdet/models/bbox_heads/double_bbox_head.py
@@ -1,9 +1,9 @@
 import torch.nn as nn
 from mmcv.cnn.weight_init import normal_init, xavier_init
 
+from mmdet.ops import ConvModule
 from ..backbones.resnet import Bottleneck
 from ..registry import HEADS
-from ..utils import ConvModule
 from .bbox_head import BBoxHead
