+++ mmdet/models/mask_heads/fcn_mask_head.py
@@ -6,10 +6,10 @@ import torch.nn as nn
 from torch.nn.modules.utils import _pair
 
 from mmdet.core import auto_fp16, force_fp32, mask_target
+from mmdet.ops import ConvModule, build_upsample_layer
 from mmdet.ops.carafe import CARAFEPack
 from ..builder import build_loss
 from ..registry import HEADS
-from ..utils import ConvModule, build_upsample_layer
 
 
 @HEADS.register_module
