+++ mmdet/models/mask_heads/fused_semantic_head.py
@@ -3,8 +3,8 @@ import torch.nn.functional as F
 from mmcv.cnn import kaiming_init
 
 from mmdet.core import auto_fp16, force_fp32
+from mmdet.ops import ConvModule
 from ..registry import HEADS
-from ..utils import ConvModule
 
 
 @HEADS.register_module
