+++ mmdet/models/mask_heads/grid_head.py
@@ -4,9 +4,9 @@ import torch.nn as nn
 import torch.nn.functional as F
 from mmcv.cnn import kaiming_init, normal_init
 
+from mmdet.ops import ConvModule
 from ..builder import build_loss
 from ..registry import HEADS
-from ..utils import ConvModule
 
 
 @HEADS.register_module
