+++ mmdet/models/mask_heads/htc_mask_head.py
@@ -1,5 +1,5 @@
+from mmdet.ops import ConvModule
 from ..registry import HEADS
-from ..utils import ConvModule
 from .fcn_mask_head import FCNMaskHead
