+++ mmdet/models/necks/fpn.py
@@ -3,8 +3,8 @@ import torch.nn.functional as F
 from mmcv.cnn import xavier_init
 
 from mmdet.core import auto_fp16
+from mmdet.ops import ConvModule
 from ..registry import NECKS
-from ..utils import ConvModule
 
 
 @NECKS.register_module
