+++ mmdet/models/necks/nas_fpn.py
@@ -2,8 +2,8 @@ import torch.nn as nn
 import torch.nn.functional as F
 from mmcv.cnn import caffe2_xavier_init
 
+from mmdet.ops import ConvModule
 from ..registry import NECKS
-from ..utils import ConvModule
 
 
 class MergingCell(nn.Module):
