--- mmdet/models/plugins/__init__.py
@@ -1,4 +1,0 @@
-from .generalized_attention import GeneralizedAttention
-from .non_local import NonLocal2D
-
-__all__ = ['NonLocal2D', 'GeneralizedAttention']
