+++ mmdet/models/utils/__init__.py
@@ -1,13 +1,3 @@
-from .conv_module import ConvModule, build_conv_layer
-from .conv_ws import ConvWS2d, conv_ws_2d
-from .norm import build_norm_layer
-from .scale import Scale
-from .upsample import build_upsample_layer
-from .weight_init import (bias_init_with_prob, kaiming_init, normal_init,
-                          uniform_init, xavier_init)
+from .weight_init import bias_init_with_prob
 
-__all__ = [
-    'conv_ws_2d', 'ConvWS2d', 'build_conv_layer', 'ConvModule',
-    'build_norm_layer', 'build_upsample_layer', 'xavier_init', 'normal_init',
-    'uniform_init', 'kaiming_init', 'bias_init_with_prob', 'Scale'
-]
+__all__ = ['bias_init_with_prob']
