+++ mmdet/ops/__init__.py
@@ -1,13 +1,21 @@
 from .context_block import ContextBlock
+from .conv import build_conv_layer
+from .conv_module import ConvModule
+from .conv_ws import ConvWS2d, conv_ws_2d
 from .dcn import (DeformConv, DeformConvPack, DeformRoIPooling,
                   DeformRoIPoolingPack, ModulatedDeformConv,
                   ModulatedDeformConvPack, ModulatedDeformRoIPoolingPack,
                   deform_conv, deform_roi_pooling, modulated_deform_conv)
+from .generalized_attention import GeneralizedAttention
 from .masked_conv import MaskedConv2d
 from .nms import nms, soft_nms
+from .non_local import NonLocal2D
+from .norm import build_norm_layer
 from .roi_align import RoIAlign, roi_align
 from .roi_pool import RoIPool, roi_pool
+from .scale import Scale
 from .sigmoid_focal_loss import SigmoidFocalLoss, sigmoid_focal_loss
+from .upsample import build_upsample_layer
 from .utils import get_compiler_version, get_compiling_cuda_version
 
 __all__ = [
@@ -16,6 +24,8 @@ __all__ = [
     'ModulatedDeformRoIPoolingPack', 'ModulatedDeformConv',
     'ModulatedDeformConvPack', 'deform_conv', 'modulated_deform_conv',
     'deform_roi_pooling', 'SigmoidFocalLoss', 'sigmoid_focal_loss',
-    'MaskedConv2d', 'ContextBlock', 'get_compiler_version',
-    'get_compiling_cuda_version'
+    'MaskedConv2d', 'ContextBlock', 'GeneralizedAttention', 'NonLocal2D',
+    'get_compiler_version', 'get_compiling_cuda_version', 'build_conv_layer',
+    'ConvModule', 'ConvWS2d', 'conv_ws_2d', 'build_norm_layer', 'Scale',
+    'build_upsample_layer'
 ]
