+++ mmdet/ops/conv_module.py
@@ -3,46 +3,9 @@ import warnings
 import torch.nn as nn
 from mmcv.cnn import constant_init, kaiming_init
 
-from mmdet.ops import DeformConvPack, ModulatedDeformConvPack
 from .activation import build_activation_layer
-from .conv_ws import ConvWS2d
+from .conv import build_conv_layer
 from .norm import build_norm_layer
-
-conv_cfg = {
-    'Conv': nn.Conv2d,
-    'ConvWS': ConvWS2d,
-    'DCN': DeformConvPack,
-    'DCNv2': ModulatedDeformConvPack,
-    # TODO: octave conv
-}
-
-
-def build_conv_layer(cfg, *args, **kwargs):
-    """ Build convolution layer
-
-    Args:
-        cfg (None or dict): cfg should contain:
-            type (str): identify conv layer type.
-            layer args: args needed to instantiate a conv layer.
-
-    Returns:
-        layer (nn.Module): created conv layer
-    """
-    if cfg is None:
-        cfg_ = dict(type='Conv')
-    else:
-        assert isinstance(cfg, dict) and 'type' in cfg
-        cfg_ = cfg.copy()
-
-    layer_type = cfg_.pop('type')
-    if layer_type not in conv_cfg:
-        raise KeyError('Unrecognized norm type {}'.format(layer_type))
-    else:
-        conv_layer = conv_cfg[layer_type]
-
-    layer = conv_layer(*args, **kwargs, **cfg_)
-
-    return layer
 
 
 class ConvModule(nn.Module):
