+++ mmdet/ops/non_local.py
@@ -2,7 +2,7 @@ import torch
 import torch.nn as nn
 from mmcv.cnn import constant_init, normal_init
 
-from ..utils import ConvModule
+from .conv_module import ConvModule
 
 
 class NonLocal2D(nn.Module):
