+++ mmdet/ops/norm.py
